2022/9/22

+++ app/src/main/java/com/example/assignment/fragment/UserProfileFragment.kt
@@ -8,20 +8,21 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.AdapterView
+import android.widget.ArrayAdapter
 import android.widget.Button
+import android.widget.ListView
 import com.example.assignment.MainActivity
 import com.example.assignment.R
 import com.example.assignment.UserLogin
+import com.example.assignment.UserProfileActivity.UserSetPersonalInfo
 
-// Below are used to do the login function
 lateinit var sharedPreferences: SharedPreferences
 //Below are the Key of login sharedPreferences
 private var PREFS_KEY = "prefs"
 private var PHONE_KEY = "phone"
 //Below are to validate the status of login
 private var phone = ""
-
-private lateinit var btnLogout:Button
 
 class UserProfileFragment : Fragment(){
 
@@ -43,14 +44,30 @@ class UserProfileFragment : Fragment(){
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         sharedPreferences = this.requireActivity().getSharedPreferences(PREFS_KEY, Context.MODE_PRIVATE)
-        btnLogout = view.findViewById(R.id.btnLogout)
-        btnLogout.setOnClickListener(){
-            val editor: SharedPreferences.Editor = sharedPreferences.edit()
-            editor.clear()
-            editor.apply()
-            val intent = Intent(this.requireActivity(), MainActivity::class.java)
-            startActivity(intent)
-            this.requireActivity().finish()
+        //List view of user profile fragment
+        val arrayAdapter: ArrayAdapter<String>
+        val userProfileOptions = arrayOf(
+            "Apply for Identity Approval", "Set Personal Information", "About Us","Logout"
+        )
+        var userProfileListView = view.findViewById<ListView>(R.id.userProfileListView)
+        arrayAdapter = ArrayAdapter(this.requireActivity(),android.R.layout.simple_list_item_1, userProfileOptions)
+        userProfileListView.adapter = arrayAdapter
+
+        userProfileListView.onItemClickListener = AdapterView.OnItemClickListener {
+                parent, view, position, id ->
+                when(position){
+                    0->{ }
+                    1->{val intent = Intent(this.requireActivity(), UserSetPersonalInfo::class.java)
+                        startActivity(intent)}
+                    2->{}
+                    3->{val editor: SharedPreferences.Editor = sharedPreferences.edit()
+                        editor.clear()
+                        editor.apply()
+                        val intent = Intent(this.requireActivity(), MainActivity::class.java)
+                        startActivity(intent)
+                        this.requireActivity().finish() }
+                }
+
         }
     }
 }
